Move unique ID helper back into sensor
